# docs/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AssetCreate": {
        "properties": {
          "asset_type": {
            "enum": [
              "domain",
              "subdomain",
              "ip"
            ],
            "title": "Asset Type",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "asset_type",
          "value"
        ],
        "title": "AssetCreate",
        "type": "object"
      },
      "AssetResponse": {
        "properties": {
          "asset_type": {
            "enum": [
              "domain",
              "subdomain",
              "ip"
            ],
            "title": "Asset Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "org_id": {
            "format": "uuid",
            "title": "Org Id",
            "type": "string"
          },
          "source": {
            "enum": [
              "user",
              "discovery"
            ],
            "title": "Source",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          },
          "verified": {
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "org_id",
          "asset_type",
          "value",
          "source",
          "verified",
          "created_at"
        ],
        "title": "AssetResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
      "NotificationResponse": {
        "properties": {
          "channel": {
            "enum": [
              "email",
              "whatsapp"
            ],
            "title": "Channel",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "org_id": {
            "format": "uuid",
            "title": "Org Id",
            "type": "string"
          },
          "recipient": {
            "title": "Recipient",
            "type": "string"
          },
          "scan_id": {
            "format": "uuid",
            "title": "Scan Id",
            "type": "string"
          },
          "sent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          },
          "status": {
            "enum": [
              "sent",
              "failed",
              "pending"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "org_id",
          "scan_id",
          "channel",
          "recipient",
          "status",
          "sent_at",
          "created_at"
        ],
        "title": "NotificationResponse",
        "type": "object"
      },
      "OrgProfileResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "domain_verified": {
            "title": "Domain Verified",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notification_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notification Email"
          },
          "primary_domain": {
            "title": "Primary Domain",
            "type": "string"
          },
          "whatsapp_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Whatsapp Number"
          }
        },
        "required": [
          "id",
          "name",
          "primary_domain",
          "whatsapp_number",
          "notification_email",
          "domain_verified",
          "created_at"
        ],
        "title": "OrgProfileResponse",
        "type": "object"
      },
      "OrgSettingsUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notification_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notification Email"
          },
          "whatsapp_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Whatsapp Number"
          }
        },
        "title": "OrgSettingsUpdate",
        "type": "object"
      },
      "PaginatedResponse_AssetResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AssetResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "title": "PaginatedResponse[AssetResponse]",
        "type": "object"
      },
...
      "PaginatedResponse_NotificationResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "title": "PaginatedResponse[NotificationResponse]",
        "type": "object"
      },
...
      "TestNotificationRequest": {
        "properties": {
          "channel": {
            "enum": [
              "email",
              "whatsapp"
            ],
            "title": "Channel",
            "type": "string"
          },
          "recipient": {
            "title": "Recipient",
            "type": "string"
          }
        },
        "required": [
          "channel",
          "recipient"
        ],
        "title": "TestNotificationRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifyTokenResponse": {
        "properties": {
          "token": {
            "title": "Token",
            "type": "string"
          },
          "txt_record": {
            "title": "Txt Record",
            "type": "string"
          }
        },
        "required": [
          "token",
          "txt_record"
        ],
        "title": "VerifyTokenResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Qelvix API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/notifications": {
      "get": {
        "description": "Log of all sent messages.",
        "operationId": "list_notifications_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_NotificationResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Notifications",
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/test": {
      "post": {
        "description": "Send test WhatsApp + email.",
        "operationId": "test_notification_notifications_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestNotificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Test Notification",
        "tags": [
          "notifications"
        ]
      }
    },
    "/org/me": {
      "delete": {
        "description": "Danger Zone: triggers DPDP erasure workflow.",
        "operationId": "delete_org_org_me_delete",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Org",
        "tags": [
          "org"
        ]
      },
      "get": {
        "description": "Get the current organization profile.",
        "operationId": "get_org_profile_org_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Org Profile",
        "tags": [
          "org"
        ]
      },
      "put": {
        "description": "Update organization settings.",
        "operationId": "update_org_settings_org_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Org Settings",
        "tags": [
          "org"
        ]
      }
    },
    "/org/me/assets": {
      "get": {
        "description": "List asset inventory.",
        "operationId": "list_assets_org_me_assets_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_AssetResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Assets",
        "tags": [
          "org"
        ]
      },
      "post": {
        "description": "Manually add an asset to monitor.",
        "operationId": "add_asset_org_me_assets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Asset",
        "tags": [
          "org"
        ]
      }
    },
    "/org/me/domain/verify-check": {
      "post": {
        "description": "Polled by Domain Verification screen to check if TXT record / file exists.",
        "operationId": "check_domain_verification_org_me_domain_verify_check_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Check Domain Verification",
        "tags": [
          "org"
        ]
      }
    },
    "/org/me/domain/verify-token": {
      "post": {
        "description": "Issues a DNS TXT / well-known-file verification token.",
        "operationId": "get_verify_token_org_me_domain_verify_token_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Verify Token",
        "tags": [
          "org"
        ]
      }
    },
...
  }
}